feat: refactor integral de vacaciones, licencias y control de asistencia

+++ src/main/java/com/indeci/rrhh/dto/PeriodoProgramadoDto.java
@@ -13,6 +13,6 @@ public record PeriodoProgramadoDto(
         Long id,
         LocalDate periodoDesde,
         LocalDate periodoHasta,
-        Integer dias,
+        Double dias,
         String tipoGoce) {
 }
